Split up DB-level unit testing
